Enable the "Details" buttons.
Right now only an empty details view is displayed.

* src/main/java/org/glom/web/client/ClientFactory.java:
* src/main/java/org/glom/web/client/ClientFactoryImpl.java:
Add DetailsView to ClientFactory.
* src/main/java/org/glom/web/client/activity/DetailsActivity.java:
A basic activity for the details view.
* src/main/java/org/glom/web/client/activity/TableSelectionActivity.java:
Add a new constructor that takes a DetailsPlace. Rename shutdown() to
clearView().
* src/main/java/org/glom/web/client/mvp/AppPlaceHistoryMapper.java:
Add DetailsPlace.Tokenizer to the list of tokens that are generated by
GWT.
* src/main/java/org/glom/web/client/mvp/DataActivityMapper.java:
Create a new DetailsActivity when a DetailsPlace is requested. Remove
unnecessary super() in constructor.
* src/main/java/org/glom/web/client/mvp/TableSelectionActivityMapper.java:
Create a new TableSelectionActivity when a DetailsPlace is requested. We
really shouldn't create a new TableSelectionActivity for both the ListPlace
and the DetailsPlace so this should be considered a temporary solution.
* src/main/java/org/glom/web/client/place/DetailsPlace.java:
New file. Represents a URL for the details view.
* src/main/java/org/glom/web/client/ui/DetailsView.java:
* src/main/java/org/glom/web/client/ui/DetailsViewImpl.java:
A basic details view interface and implementation.
* src/main/java/org/glom/web/client/ui/ListViewImpl.java:
Enable the "Details" buttons.