Display main layout group titles in the DetailsView.
This is the start of the DetailsActivity/DetailsView implementation.

* src/main/java/org/glom/web/client/OnlineGlomService.java:
* src/main/java/org/glom/web/client/OnlineGlomServiceAsync.java:
Add getDetailsLayoutGroup() and getDefaultDetailsLayoutGroup() methods.
* src/main/java/org/glom/web/client/activity/DetailsActivity.java:
Get the layout information for the details view from the server and set
the main layout group titles.
* src/main/java/org/glom/web/client/ui/DetailsView.java:
* src/main/java/org/glom/web/client/ui/DetailsViewImpl.java:
Add addLayoutGroup() and addLayoutField() methods. This are just
temporary methods for creating the the details view that will change
in the future.
* src/main/java/org/glom/web/server/OnlineGlomServiceImpl.java:
Implement getDetailsLayoutGroup() and getDefaultDetailsLayoutGroup()
methods.
* src/main/java/org/glom/web/shared/layout/Formatting.java:
* src/main/java/org/glom/web/shared/layout/LayoutGroup.java:
* src/main/java/org/glom/web/shared/layout/LayoutItem.java:
* src/main/java/org/glom/web/shared/layout/LayoutItemField.java:
* src/main/java/org/glom/web/shared/layout/LayoutItemWithFormatting.java:
Data Transfer Objects that mimic the libglom object structure. These are
used for transferring the details layout but could also be used for
transferring the list layout.