Add data to the DetailsView using a hard-coded primary key value.
The layout and functionality of the DetailsView is not complete. This
is just a checkpoint so the patch doesn't get too big.

* src/main/java/org/glom/web/client/OnlineGlomService.java:
* src/main/java/org/glom/web/client/OnlineGlomServiceAsync.java:
Add getDetailsData() servlet method.
* src/main/java/org/glom/web/client/activity/DetailsActivity.java:
Add RPC to getDetailsData(). Change the way the LayoutGroups and
LayoutFields are added to the DetailsView.
* src/main/java/org/glom/web/client/ui/DetailsView.java:
* src/main/java/org/glom/web/client/ui/DetailsViewImpl.java:
Add setData() method. Change addLayoutGroup() and addLayoutField() to
take the string for the title instead of the object.
* src/main/java/org/glom/web/server/OnlineGlomServiceImpl.java:
Add implementation getDetailsData() along with some private helper
methods.
* src/main/webapp/style.css: Add padding to details-data class. Add a
details-label class with the same padding as the details-data class.