Use the LayoutGroup group for the list layout instead of ColumnInfo and LayoutListTable.
This makes things more inline with how libglom works and reduces code
duplication. This refactor lays the groundwork for adding the primary key to
the LayoutGroup object.

* src/main/java/org/glom/web/client/OnlineGlomService.java:
* src/main/java/org/glom/web/client/OnlineGlomServiceAsync.java:
* src/main/java/org/glom/web/server/OnlineGlomServiceImpl.java:
Change method names to getListLayout and getDefaultListLayout for
consistency. Use LayoutGroup as the DTO for the list layout instead of
ColumnInfo and LayoutListTable.
* src/main/java/org/glom/web/client/activity/ListActivity.java: Use the
new method names along with the LayoutGroup object for transferring the
list layout.
* src/main/java/org/glom/web/client/ui/ListView.java:
* src/main/java/org/glom/web/client/ui/ListViewImpl.java:
Use ArrayList<LayoutItem> instead of ColumnInfo[] in setCellTable().
* src/main/java/org/glom/web/shared/ColumnInfo.java: Deleted. Replaced
with LayoutGroup.
* src/main/java/org/glom/web/shared/LayoutListTable.java: Deleted.
Replaced with LayoutGroup.
* src/main/java/org/glom/web/shared/layout/LayoutGroup.java: Add
expectedResultSize and defaultTableName fields which are needed for
the list layout.
* src/main/java/org/glom/web/shared/layout/LayoutItemField.java: Add
type field which is needed for the list layout but will also be
useful for the details layout as things progress.
* src/main/java/org/glom/web/shared/layout/LayoutItemWithFormatting.java:
Make class abstract. Remove the unnecessary
getFormattingHorizontalAlignment method. Add setFormatting method.