Set the height of the data element in the Details View.
I changed the InlineLabels (text in a span element) to Labels (text in
a div element) so that I could set the height of the details-data
elements instead of the details-cell parent elements. This allows the
the details-data element to display the correct height if style is
applied that shows the height.

This change has the added benefit of allowing the order of the labels
and data elements to be changed for right-to-left languages.

* src/main/java/org/glom/web/client/ui/DetailsViewImpl.java: Change
InlineLabels to Labels.
* src/main/java/org/glom/web/client/ui/FlowTable.java: Change
InlineLabels to Labels. Set the height of the data element.
* src/main/java/org/glom/web/server/OnlineGlomServiceImpl.java: Set the
multiline text height in the Formatting DTO.
* src/main/java/org/glom/web/shared/layout/Formatting.java: Add field
for multiline height along with getter and setter methods.
* src/main/webapp/style.css: Adjust style to account for the change
from span elements to div elements in the details cell.