Remove GWT styling from open button in details view.
There are still some issues with how the details cell is arranged but
this should be made to match Glom 1.20. I'm going to leave fixing this
until I have Glom 1.20 up and running.

* src/main/java/org/glom/web/client/ui/details/DetailsCell.java: Set
style name on open button.
* src/main/webapp/style.css: Move and edit details-navigation class.
Re-arrange some classes to make them appear in the same order as the
UI.