Create Notebook widgets to the details view.
This isn't finished just yet - I still need to create a reasonable
style to match the current theme.

* src/main/java/org/glom/web/client/Utils.java: Add method for
calculating the height of a widget. This is used in the Notebook class.
* src/main/java/org/glom/web/client/ui/DetailsViewImpl.java: Update for
new constructor method in Group.
* src/main/java/org/glom/web/client/ui/details/Group.java: Add new
method for creating child widget that can be used by subclasses
like Notebook. New constructor that allows disabling the group
titles - Notebooks don't set a group title for their child groups.
* src/main/java/org/glom/web/client/ui/details/Notebook.java: New class
to make Notebooks using GWT's TabLayoutPanel.
* src/main/java/org/glom/web/client/ui/details/Portal.java: New
constructor that allows disabling the group titles.
* src/main/java/org/glom/web/server/ConfiguredDocument.java: Create the
LayoutItemNotebook DTO.
* src/main/java/org/glom/web/shared/layout/LayoutItemNotebook.java: New
DTO for Notebooks. It's just an empty class for now but we might need
it to transfer some specific information in the future.