Ensure the pager buttons are always visible for related lists.
To accomplish this, I've turned off text wrapping in the list view and
related list tables for both the header and data text. The related list
table now has a fixed layout so the it doesn't overflow its container.
This is required to ensure that the cell text is clipped when it
overflows the cell and an ellipsis is added to the right side of the
cell when text is clipped.

A fixed table layout for the related list table in the details view
seems what we want for the details view anyway, so the side-effect is
desirable.

The ellipsis will only be displayed in Firefox >= 7.

This fixes bug:

https://bugzilla.gnome.org/show_bug.cgi?id=662930

* src/main/java/org/glom/web/client/ui/cell/NumericCell.java:
* src/main/java/org/glom/web/client/ui/cell/TextCell.java: Add
'overflow: hidden; text-overflow: ellipsis;' CSS properties to the table
cell text.
* src/main/java/org/glom/web/client/ui/details/RelatedListTable.java:
Set the 'table-layout: fixed' CSS property to the related list table.
* src/main/java/org/glom/web/client/ui/list/ListTable.java: Set the
'white-space: nowrap;' CSS property on both the list view and the
related list tables.