Improve browser configuration error messages.
This fixes:

https://bugzilla.gnome.org/show_bug.cgi?id=662792

* src/main/java/org/glom/web/client/OnlineGlomService.java:
* src/main/java/org/glom/web/client/OnlineGlomServiceAsync.java: Add
getConfigurationErrorMessage() method.
* src/main/java/org/glom/web/client/activity/DocumentSelectionActivity.java:
Get and display a specific configuration error message when no Glom
documents are found.
* src/main/java/org/glom/web/server/OnlineGlomServiceImpl.java:
Implement getConfigurationErrorMessage() method. Surround configuration
code in the init() method with a try/catch block. This allows the
errors to be caught while keeping the servlet available to retrieve the
configuration error message.